Make it work for CounterList.
